requirements.txt
aniso8601==9.0.1
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.4
django-multiselectfield==0.1.12
djangorestframework==3.14.0
graphene==3.1.1
graphene-django==3.0.0
graphene-graphiql-explorer==0.1.1
graphql-core==3.2.3
graphql-relay==3.2.0
promise==2.3
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
tzdata==2022.7
